int64 indexing: comprehensive migration progress
Core Types:
- Shape.cs, Shape.Reshaping.cs: int to long for dimensions, strides, offsets
- ArraySlice.cs, ArraySlice`1.cs: int64 indexing support
- UnmanagedStorage.cs, Getters, Setters: long indices throughout
- UnmanagedMemoryBlock`1.cs, Casting.cs: memory block int64 support
- Slice.cs: slice indices as long

Iterators:
- NDIterator.cs, MultiIterator.cs: long coordinate support
- All Incrementors: NDCoordinatesIncrementor, AxisIncrementor, etc.

IL Kernels:
- ILKernelGenerator (Binary, Unary, Comparison, Reduction, Scan, Shift, Clip, Modf, Masking): long offsets
- KernelSignatures.cs, BinaryKernel.cs, ReductionKernel.cs: delegate signatures updated
- SimdReductionOptimized.cs, SimdThresholds.cs: SIMD path updates

DefaultEngine:
- All reduction operations (Add, AMax, AMin, ArgMax, ArgMin, Mean, Std, Var, etc.)
- BinaryOp, UnaryOp, CompareOp: long indices
- Dot.NDMD, MatMul.2D2D, Clip, Modf, Shift: BLAS operations updated

NDArray & APIs:
- NDArray.cs, NDArray.Unmanaged.cs, NDArray.String.cs: core array updates
- TensorEngine.cs, NDArray`1.cs: engine interface updates
- Indexing (Selection, Masking): long index support
- np.size, np.array, np.zeros, np.ones, np.full, np.meshgrid, etc.

New files:
- np.count_nonzero.cs: new API implementation
- Default.IsInf.cs: infinity checking
- CountNonzeroTests.cs: test coverage

Tests updated across all affected components.

Part of int64 indexing migration (#584)

+++ src/NumSharp.Bitmap/np_.extensions.cs
@@ -230,7 +230,7 @@ namespace NumSharp
             if (nd.shape[0] != 1)
                 throw new ArgumentException($"ndarray has more than one picture in it ({nd.shape[0]}) based on the first dimension.");
 
-            var bbp = nd.shape[3]; //bytes per pixel.
+            var bbp = (int)nd.shape[3]; //bytes per pixel.
             if (bbp != extractFormatNumber())
                 throw new ArgumentException($"Given PixelFormat: {format} does not match the number of bytes per pixel in the 4th dimension of given ndarray.");
 
@@ -284,7 +284,7 @@ namespace NumSharp
                             break;
                     }
 
-                    return (int)bbp;
+                    return bbp;
                 }
 
                 return format.ToBytesPerPixel();
@@ -307,10 +307,8 @@ namespace NumSharp
             if (nd.shape[0] != 1)
                 throw new ArgumentException($"ndarray has more than one picture in it ({nd.shape[0]}) based on the first dimension.");
 
-            var height = nd.shape[1];
-            var width = nd.shape[2];
-            
-            //TODO: if bigger than int.maxvalue then throw execption about bitmap supporting only up to int.maxvalue.
+            var height = (int)nd.shape[1];
+            var width = (int)nd.shape[2];
 
             return ToBitmap(nd, width, height, format);
         }
